requirements.txt
# ─── Core Framework ───
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.8.1

# ─── Supabase ───
supabase==2.13.0
postgrest==0.19.1

# ─── Auth / JWT ───
python-jose[cryptography]==3.4.0
PyJWT==2.10.1

# ─── AI: Gemini ───
google-genai==1.12.1

# ─── AI: Whisper (local STT) ───
openai-whisper==20240930

# ─── TTS: Google Translate ───
gTTS==2.5.4

# ─── HTTP Client ───
httpx==0.28.1

# ─── PDF / Image Processing ───
Pillow==11.2.1
PyPDF2==3.0.1
pdf2image==1.17.0

# ─── Environment ───
python-dotenv==1.1.0

# ─── Utilities ───
uuid7==0.1.0
